refactor: move RhoType to his own file
